notifications-service/worker/pyproject.toml
[tool.poetry]
name = "worker"
version = "0.1.0"
description = ""
authors = ["s.novik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aio-pika = "^9.3.0"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
